refactor(session): Used built-in mmkv hook for getting storage value
