+++ src/entities/session/model/hooks/useHasSession.ts
@@ -1,7 +1,7 @@
 import useSession from './useSession';
 
 function useHasSession() {
-  return !!useSession().accessToken;
+  return !!useSession()?.accessToken;
 }
 
 export default useHasSession;
